Add files via upload
By changing the classifier threshold we can change the precision and recall depending on the demand of the business. Here we were able to improve the recall at the expense of precision, but for business, recall is more important. Say if we want to improve the churn by giving a discount on charges and other promotions, it's okay to give customers these offers that we thought would churn. In this way at least we can try to stop the actual churn.